+++ src/main/java/com/example/service/EntrenadorServiceImp.java
@@ -51,8 +51,11 @@ public class EntrenadorServiceImp implements IEntrenadorService {
         Optional<Entrenador> entrenador = Optional.ofNullable(this.entrenadorRepository.find(nombre)
                 .orElseThrow(() -> new NotFoundException("El entrenador " + nombre + " no fue encontrado.")));
 
-        this.boxeadorRepository.eliminarEntrenador(entrenador.get());
-        return this.entrenadorRepository.delete(entrenador.get());
+        if(entrenador.isPresent()) {
+            this.boxeadorRepository.eliminarEntrenador(entrenador.get());
+            return this.entrenadorRepository.delete(entrenador.get());
+        }
+        return false;
     }
 
     @Override
@@ -60,9 +63,11 @@ public class EntrenadorServiceImp implements IEntrenadorService {
         Optional<Entrenador> entrenador = Optional.ofNullable(this.entrenadorRepository.find(nombre)
                 .orElseThrow(() -> new NotFoundException("El entrenador " + nombre + " no fue encontrado.")));
 
-        entrenador.get().setNombre(entrenadorDto.getNombre());
-        entrenador.get().setCategorias(entrenadorDto.getCategorias());
-        entrenador.get().setBoxeadores(entrenadorDto.getBoxeadores().stream().map(b -> modelMapper.map(b, Boxeador.class)).collect(Collectors.toList()));
+        if(entrenador.isPresent()) {
+            entrenador.get().setNombre(entrenadorDto.getNombre());
+            entrenador.get().setCategorias(entrenadorDto.getCategorias());
+            entrenador.get().setBoxeadores(entrenadorDto.getBoxeadores().stream().map(b -> modelMapper.map(b, Boxeador.class)).collect(Collectors.toList()));
+        }
 
         return Optional.of(modelMapper.map(entrenador, EntrenadorDto.class));
     }
